Add various simplifications and more robust GitHub actions for the check-in test (#225)
Minor GitHub action changes. The biggest one is that we have verified that the Gradle caches work between steps, branches and jobs, so nothing building close in time cannot overwrite the GitHub workflow caches for another commit.

We also make sure that all installDist/manualTest verification for the test platforms is finished before trying to publish a snapshot artifact. This is to make sure that even if we run on the publishing platform (ubuntu-latest), and windows fails, we do not publish anything that may be a windows-breaker.

This is all good prep work for the jreleaser configuration which will accept an uploaded build artifact blob consisting of the installWithLaunchersDist destination directories after uploadArtifact.

We have also added various cleaners that are meant to run either manually dispatched, and on a cron schedule, so that we keep the number of published snapshots low, with only the ~10 latest ones for the same version. Furthermore we may want to delete old caches for old runs, as well as workflow results, when they cease to be relevant.