Reenable homebrew tap idk-latest again, and update every time something is merged into master (finally) (#293)
Note that the various artifact build steps can always be trigged by manual dispatch if you are not in master and need to test them. 

We now build and update snapshot releases, and installDist defaults to creating launcher scripts for xec and xcc. We can still create binary launchers with ./gradlew xdk:installWithLaunchersDist, like we always can. The default docker and maven and Github snapshot releases default to using installDist with scripts for xec and xcc, which we can both improve and change later if we don't like it. Or compile them inte native exes or whatever.